+++ includes/api/cocart/v1/class-cocart-controller.php
@@ -6,7 +6,7 @@
  *
  * @author   Sébastien Dumont
  * @category API
- * @package  CoCart/API/v1
+ * @package  CoCart\API\v1
  * @since    2.0.0
  * @version  2.5.0
  * @license  GPL-2.0+