+++ includes/class-cocart-autoloader.php
@@ -4,7 +4,7 @@
  *
  * @author   Sébastien Dumont
  * @category Classes
- * @package  CoCart/Classes/AutoLoader
+ * @package  CoCart\Classes\AutoLoader
  * @since    2.0.0
  * @license  GPL-2.0+
  */