+++ includes/class-cocart-server.php
@@ -6,7 +6,7 @@
  *
  * @author   Sébastien Dumont
  * @category API
- * @package  CoCart/API
+ * @package  CoCart\API
  * @since    1.0.0
  * @version  3.0.0
  * @license  GPL-2.0+
@@ -44,9 +44,6 @@ class CoCart_Server {
 
 		// Hook into WordPress ready to init the REST API as needed.
 		add_action( 'rest_api_init', array( $this, 'register_rest_routes' ), 10 );
-
-		// Allow all cross origin requests.
-		add_action( 'rest_api_init', array( $this, 'allow_all_cors' ), 15 );
 	} // END __construct()
 
 	/**
@@ -138,28 +135,30 @@ class CoCart_Server {
 	/**
 	 * Loads the cart, session and notices should it be required.
 	 * 
-	 * Note: Only needed should the site be running WooCommerce 3.6 
-	 * or higher as they are not included during a REST request.
-	 *
 	 * @access  private
 	 * @since   2.0.0
-	 * @version 2.3.0
+	 * @version 2.6.0
 	 */
 	private function maybe_load_cart() {
-		if ( CoCart_Helpers::is_wc_version_gte_3_6() && CoCart_Helpers::is_rest_api_request() ) {
-			require_once( WC_ABSPATH . 'includes/wc-cart-functions.php' );
-			require_once( WC_ABSPATH . 'includes/wc-notice-functions.php' );
+		if ( CoCart_Helpers::is_rest_api_request() ) {
+			// WooCommerce is greater than v3.6 or less than v4.5
+			if ( CoCart_Helpers::is_wc_version_gte_3_6() && CoCart_Helpers::is_wc_version_lt_4_5() ) {
+				require_once( WC_ABSPATH . 'includes/wc-cart-functions.php' );
+				require_once( WC_ABSPATH . 'includes/wc-notice-functions.php' );
 
-			// Disable cookie authentication REST check and ONLY, if the site is SECURE. 🔒
-			if ( is_ssl() ) {
-				remove_filter( 'rest_authentication_errors', 'rest_cookie_check_errors', 100 );
+				// Initialize session.
+				$this->initialize_session();
+
+				// Initialize cart.
+				$this->initialize_cart();
 			}
 
-			// Initialize session.
-			$this->initialize_session();
-
-			// Initialize cart.
-			$this->initialize_cart();
+			// WooCommerce is greater than v4.5 or equal.
+			if ( CoCart_Helpers::is_wc_version_gte_4_5() ) {
+				if ( is_null( WC()->cart ) && function_exists( 'wc_load_cart') ) {
+					wc_load_cart();
+				}
+			}
 
 			// Identify if user has switched.
 			if ( $this->has_user_switched() ) {
@@ -265,8 +264,6 @@ class CoCart_Server {
 	 * @version 2.5.0
 	 */
 	public function rest_api_includes() {
-		$this->maybe_load_cart();
-
 		// Only include Legacy REST API if WordPress is v5.4.2 or lower.
 		if ( CoCart_Helpers::is_wp_version_lt( '5.4.2' ) ) {
 			// Legacy - WC Cart REST API v2 controller.
@@ -298,54 +295,6 @@ class CoCart_Server {
 
 		do_action( 'cocart_rest_api_controllers' );
 	} // rest_api_includes()
-
-	/**
-	 * Allow all cross origin header requests.
-	 * 
-	 * Disabled by default. Requires `cocart_allow_all_cors` filter set to true to enable.
-	 *
-	 * @access  public
-	 * @since   2.2.0
-	 * @version 2.3.0
-	 */
-	public function allow_all_cors() {
-		// If not enabled via filter then return.
-		if ( apply_filters( 'cocart_disable_all_cors', true ) ) {
-			return;
-		}
-
-		// If the REST API request was not for CoCart then return.
-		if ( ! CoCart_Helpers::is_rest_api_request() ) {
-			return;
-		}
-
-		// Remove the default cors server headers.
-		remove_filter( 'rest_pre_serve_request', 'rest_send_cors_headers' );
-
-		// Adds new cors server headers.
-		add_filter( 'rest_pre_serve_request', array( $this, 'cors_headers' ), 0, 4 );
-	} // END allow_all_cors()
-
-	/**
-	 * Cross Origin headers.
-	 *
-	 * @access  public
-	 * @since   2.2.0
-	 * @version 2.5.1
-	 * @param   bool             $served  Whether the request has already been served. Default false.
-	 * @param   WP_HTTP_Response $result  Result to send to the client. Usually a WP_REST_Response.
-	 * @param   WP_REST_Request  $request Request used to generate the response.
-	 * @param   WP_REST_Server   $server  Server instance.
-	 * @return  bool
-	 */
-	public function cors_headers( $served, $result, $request, $server ) {
-		header( 'Access-Control-Allow-Origin: ' . apply_filters( 'cocart_allow_origin', $_SERVER['HTTP_ORIGIN'] ) );
-		header( 'Access-Control-Allow-Methods: POST, GET, OPTIONS, DELETE' );
-		header( 'Access-Control-Allow-Credentials: true' );
-		header( 'Access-Control-Allow-Headers: Authorization, Content-Type, X-Requested-With' );
-
-		return $served;
-	} // END cors_headers()
 
 } // END class