+++ CHANGELOG.md
@@ -27,6 +27,19 @@ Updating to the new API is easy. All parameters are the same. Simply follow the 
 * * `cocart_item_price`
 * * `cocart_item_subtotal`
 
+## v2.6.0 - 8th September, 2020
+
+* **NEW**: Added backwards compatibility for when `set_customer_session_cookie()` function is called.
+* **NEW**: Site environment is now checked before plugin activates.
+* **NEW**: Plugin will not activate if CoCart Pro _v1.1.0_ or above is installed.
+* Enhanced: Plugin to be better optimized for future releases of WooCommerce and to allow it to be packaged with CoCart Pro so user's will not require CoCart Lite if Pro is installed.
+* Improved: `uninstall.php` file and now will only clear plugin data and remove the database table if `COCART_REMOVE_ALL_DATA` constant is set to true in user's `wp-config.php`. This is to prevent data loss when deleting the plugin from the backend and to ensure only the site owner can perform this action.
+* Improved: Handling of admin menu by moving to it's own class.
+* Tweaked: Admin notices to not redirect only to the plugins page once notice is dismissed. Will redirect to the current admin page.
+* Dev: Introduced the ability to white label the plugin. Meaning hide **CoCart** from the backend. (Admin menu, plugin links, plugin notices including WC admin inbox notices and WC system status information). All you have to do is set a constant `COCART_WHITE_LABEL` to true in user's `wp-config.php` file.
+* Dev: New filter `cocart_authenticate` to override the determined user ID when authenticating. **NOTE** This will only be active if not already authenticated.
+* Tested: Compatible with WooCommerce v4.5.0
+
 ## v2.5.1 - 18th August, 2020
 
 * Fixed: **'Access-Control-Allow-Origin'** header response when it request's credentials with a wildcard `(*)` value.